requirements.txt
requests>=2.31.0
PyYAML>=6.0.1
python-dotenv>=1.0.1
tqdm>=4.66.4
langdetect>=1.0.9
transformers>=4.41.0
sentencepiece>=0.1.99
torch>=2.2.0
huggingface-hub>=0.24.0
